Rename package to be shorter
