+++ examples/rabbitmq-auth-backend-spring-boot-kotlin/src/main/kotlin/com/rabbitmq/examples/RabbitmqAuthBackendSpringBootKotlinApplication.kt
@@ -1,4 +1,4 @@
-package com.rabbitmq.examples.rabbitmqauthbackendspringbootkotlin
+package com.rabbitmq.examples
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
