+++ examples/rabbitmq-auth-backend-spring-boot-kotlin/src/test/kotlin/com/rabbitmq/examples/RabbitmqAuthBackendSpringBootKotlinApplicationTests.kt
@@ -1,4 +1,4 @@
-package com.rabbitmq.examples.rabbitmqauthbackendspringbootkotlin
+package com.rabbitmq.examples
 
 import org.junit.Test
 import org.junit.runner.RunWith
